Add the plugin-ified fedup2
It turns out we can do fedup2 almost entirely as a DNF plugin!

Really the only thing we're missing is the ability for plugin commands
to set/modify the TransactionDisplay object that gets used to display
the progress of the upgrade transaction.

Here's the prototype plugin version, uploaded so I don't lose it.